pyproject.toml
[tool.poetry]
name = "re-data-engineering"
version = "0.1.0"
description = "Automated systems to gather data from leading online real estate portals and various other resources leveraging cloud solutions."
authors = ["Anderson Arroyo <a.arroyo@pucp.pe>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.8.4"
beautifulsoup4 = "4.12.2"
geopy = "2.3.0"
google-cloud-storage = "2.8.0"
google-cloud-bigquery = "3.10.0"
httpx = "0.24.0"
pandas = "2.0.0"
prefect = "2.10.6"
prefect-gcp = "0.4.1"
pyarrow = "11.0.0"
tqdm = "4.65.0"
latest-user-agents = "0.0.3"
httpagentparser = "1.9.5"
pydantic = "1.10.11"
anyio = "3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
